requirements.txt
python-dotenv==1.0.1
google-ai-generativelanguage==0.4.0
